# Better text output encoding