--- a/encoding/buffer/buffer.go
+++ b/encoding/buffer/buffer.go
@@ -1,4 +1,4 @@
-package json
+package buffer
 
 import "strconv"