--- a/encoding/buffer/buffer_suite_test.go
+++ b/encoding/buffer/buffer_suite_test.go
@@ -1,0 +1,13 @@
+package buffer_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestBuffer(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Buffer Suite")
+}