--- a/encoding/text/text_suite_test.go
+++ b/encoding/text/text_suite_test.go
@@ -1,0 +1,13 @@
+package text_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestText(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Text Suite")
+}