--- a/encoding/types/array.go
+++ b/encoding/types/array.go
@@ -1,0 +1,6 @@
+package types
+
+// Array is a JSON marshalable array
+type Array interface {
+	MarshalArray(enc Encoder) error
+}