--- a/entry.go
+++ b/entry.go
@@ -7,6 +7,7 @@ import (
 
 	json "github.com/autopilothq/lg/encoding"
 	fancy "github.com/autopilothq/lg/encoding/json"
+	text "github.com/autopilothq/lg/encoding/text"
 )
 
 // Entry represents a log entry
@@ -19,14 +20,46 @@ type Entry struct {
 
 const (
 	// TimeFormat is the layout to use when rendering time
-	TimeFormat = "2006-01-02T15:04:05.999"
+	TimeFormat = "2006-01-02T15:04:05.000"
 )
 
 func (e *Entry) toPlainText() []byte {
-	timeBytes := bytes.NewBufferString(e.Timestamp.Format(TimeFormat)).Bytes()
-	return append(
-		append(timeBytes, ' '),
-		e.toPlainTextWithoutTime()...)
+	timeBytes := bytes.NewBufferString(e.Timestamp.Format(TimeFormat))
+
+	_, err := timeBytes.WriteString(e.Level.AlignedString())
+	if err != nil {
+		return append([]byte(err.Error()), '\n')
+	}
+
+	if e.Fields.Len() > 0 {
+		enc := text.NewEncoder()
+
+		if err = enc.StartArray(); err != nil {
+			return append([]byte(err.Error()), '\n')
+		}
+
+		if err = e.Fields.encodeText(enc); err != nil {
+			return append([]byte(err.Error()), '\n')
+		}
+
+		if err = enc.EndArray(); err != nil {
+			return append([]byte(err.Error()), '\n')
+		}
+
+		if _, err = timeBytes.Write(enc.Bytes()); err != nil {
+			return append([]byte(err.Error()), '\n')
+		}
+
+		if err = timeBytes.WriteByte(' '); err != nil {
+			return append([]byte(err.Error()), '\n')
+		}
+	}
+
+	if _, err = timeBytes.WriteString(e.Message); err != nil {
+		return append([]byte(err.Error()), '\n')
+	}
+
+	return append(timeBytes.Bytes(), '\n')
 }
 
 func (e *Entry) toPlainTextWithoutTime() []byte {
@@ -38,7 +71,7 @@ func (e *Entry) toPlainTextWithoutTime() []byte {
 func makeJSONError(enc *fancy.Encoder, err error) []byte {
 	b := fmt.Sprintf(`{"error": "encoding error: %s"}`,
 		json.EncodeValue(enc, err))
-	return []byte(b)
+	return append([]byte(b), '\n')
 }
 
 func (e *Entry) toJSON() []byte {