--- a/fields.go
+++ b/fields.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/autopilothq/lg/encoding"
 	fancy "github.com/autopilothq/lg/encoding/json"
+	text "github.com/autopilothq/lg/encoding/text"
 )
 
 // F represents a single pair of log 'fields'
@@ -16,6 +17,11 @@ type F struct {
 // Fields is a slice of field key-value pairs
 type Fields struct {
 	contents []F
+}
+
+// Len returns the number of fields
+func (f *Fields) Len() int {
+	return len(f.contents)
 }
 
 func (f *Fields) renderPlainText() string {
@@ -61,12 +67,30 @@ func (f *Fields) set(fld F) {
 
 // encodeJSON allows Fields to be marshaled to JSON via the encoder
 func (f *Fields) encodeJSON(encoder *fancy.Encoder) (err error) {
-	if len(f.contents) > 0 {
-		for _, fld := range f.contents {
-			err := encoding.EncodeKeyValue(encoder, fld.Key, fld.Val)
-			if err != nil {
-				return err
-			}
+	if len(f.contents) == 0 {
+		return nil
+	}
+
+	for _, fld := range f.contents {
+		err := encoding.EncodeKeyValue(encoder, fld.Key, fld.Val)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// encodeText allows Fields to be marshaled to TEXT via the encoder
+func (f *Fields) encodeText(encoder *text.Encoder) (err error) {
+	if len(f.contents) == 0 {
+		return nil
+	}
+
+	for _, fld := range f.contents {
+		err := encoding.EncodeKeyValue(encoder, fld.Key, fld.Val)
+		if err != nil {
+			return err
 		}
 	}