--- a/levels.go
+++ b/levels.go
@@ -38,6 +38,26 @@ func (l Level) String() string {
 	}
 }
 
+// AlignedString returns the string form of the Level aligned for output
+func (l Level) AlignedString() string {
+	switch l {
+	case LevelTrace:
+		return " trace "
+	case LevelDebug:
+		return " debug "
+	case LevelInfo:
+		return " info  "
+	case LevelWarn:
+		return " warn  "
+	case LevelError:
+		return " error "
+	case LevelFatal:
+		return " fatal "
+	default:
+		return "       "
+	}
+}
+
 func (l Level) MarshalJSON() ([]byte, error) {
 	return json.Marshal(l.String())
 }