This is the sister change to the previous JSON one

* decrease memory allocations of the text output
* decrease amount of memory allocated by text output
* increase throughput of text output
* some bits of json encoding are now shared (buffer, buffer pooling, collection marshalling)

The following benchmark compares the old reflection based encoding to the new
version for both text and json

```
benchmark                                                        old ns/op     new ns/op     delta
BenchmarkFieldLogging/Marshalling_JSON-8                         32718         6133          -81.25%
BenchmarkFieldLogging/Marshalling_JSON_parallel-8                10322         2878          -72.12%
BenchmarkPlainTextFieldLogging/Marshalling_Fields-8              22254         6625          -70.23%
BenchmarkPlainTextFieldLogging/Marshalling_Fields_parallel-8     6841          2518          -63.19%

benchmark                                                        old allocs     new allocs     delta
BenchmarkFieldLogging/Marshalling_JSON-8                         166            49             -70.48%
BenchmarkFieldLogging/Marshalling_JSON_parallel-8                166            49             -70.48%
BenchmarkPlainTextFieldLogging/Marshalling_Fields-8              163            53             -67.48%
BenchmarkPlainTextFieldLogging/Marshalling_Fields_parallel-8     163            53             -67.48%

benchmark                                                        old bytes     new bytes     delta
BenchmarkFieldLogging/Marshalling_JSON-8                         25459         5028          -80.25%
BenchmarkFieldLogging/Marshalling_JSON_parallel-8                47192         18704         -60.37%
BenchmarkPlainTextFieldLogging/Marshalling_Fields-8              18679         5537          -70.36%
BenchmarkPlainTextFieldLogging/Marshalling_Fields_parallel-8     36307         17565         -51.62%
```