@luma 2017-12-20
I'm implemented a AddTime encoding helper. It's faster than using time.Format but still slower than I'd like. It's probably worth switching low times to use a nano-precision timestamp to avoid taking a hit when output the time of every log line.

```
benchmark                                                        old ns/op     new ns/op     delta
BenchmarkFieldLogging/Marshalling_JSON-8                         32718         8510          -73.99%
BenchmarkFieldLogging/Marshalling_JSON_parallel-8                10322         4190          -59.41%
BenchmarkPlainTextFieldLogging/Marshalling_Fields-8              22254         8311          -62.65%
BenchmarkPlainTextFieldLogging/Marshalling_Fields_parallel-8     6841          3266          -52.26%

benchmark                                                        old allocs     new allocs     delta
BenchmarkFieldLogging/Marshalling_JSON-8                         166            49             -70.48%
BenchmarkFieldLogging/Marshalling_JSON_parallel-8                166            49             -70.48%
BenchmarkPlainTextFieldLogging/Marshalling_Fields-8              163            53             -67.48%
BenchmarkPlainTextFieldLogging/Marshalling_Fields_parallel-8     163            53             -67.48%

benchmark                                                        old bytes     new bytes     delta
BenchmarkFieldLogging/Marshalling_JSON-8                         25459         5463          -78.54%
BenchmarkFieldLogging/Marshalling_JSON_parallel-8                47192         24086         -48.96%
BenchmarkPlainTextFieldLogging/Marshalling_Fields-8              18679         6016          -67.79%
BenchmarkPlainTextFieldLogging/Marshalling_Fields_parallel-8     36307         22130         -39.05%
```