@luma 2017-12-20
Gonna be cheeky and merge this now. I've been testing it with the banks integration tests all day :-)